Fixed styling

+++ src/main/java/me/j4n8/diplomskanaloga/frontend/components/projects/ProjectMemberComponent.java
@@ -1,8 +1,10 @@
 package me.j4n8.diplomskanaloga.frontend.components.projects;
 
 import com.vaadin.flow.component.button.Button;
+import com.vaadin.flow.component.confirmdialog.ConfirmDialog;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.Span;
+import com.vaadin.flow.theme.lumo.LumoUtility;
 import me.j4n8.diplomskanaloga.project.ProjectService;
 import me.j4n8.diplomskanaloga.project.entities.ProjectEntity;
 import me.j4n8.diplomskanaloga.user.entities.UserEntity;
@@ -10,8 +12,7 @@ import me.j4n8.diplomskanaloga.user.entities.UserEntity;
 public class ProjectMemberComponent extends Div {
 	private final UserEntity member;
 	private final ProjectEntity project;
-	private Span username;
-	private Span email;
+	private Span details;
 	private Button removeButton;
 	private final ProjectService projectService;
 	
@@ -19,18 +20,27 @@ public class ProjectMemberComponent extends Div {
 		this.projectService = projectService;
 		this.member = member;
 		this.project = project;
-		
-		username = new Span(member.getUsername());
-		email = new Span(member.getEmail());
+		details = new Span(member.getUsername() + " | " + member.getEmail());
 		removeButton = new Button("Remove");
 		removeButton.addClickListener(event -> {
 			removeMember();
 		});
 		
-		add(username, email, removeButton);
+		add(details, removeButton);
+		
+		applyStyles();
+	}
+	
+	private void applyStyles() {
+		addClassNames(LumoUtility.Display.FLEX, LumoUtility.Gap.SMALL);
+		details.setWidthFull();
 	}
 	
 	private void removeMember() {
-		projectService.removeMember(member, project);
+		new ConfirmDialog("Are you sure you want to remove this member?", "This action cannot be undone",
+				"Remove", e -> {
+			projectService.removeMember(member, project);
+		}, "Cancel", e -> {
+		}).open();
 	}
 }
